Swap webthing instances to run as processes
Webthing instances were previously implemented as threads, which
limited how hard they could exercise the gateway (tasks quickly became
CPU bound).  This change modifies the testThing class so that it spawns
separate user processes.

Additionally, this commit modifies how the wait for webthings to be ready
is implemented - less code, more good.

When running against a localhost gateway, the gateway run command can
be prefixed with 'taskset 0x1 <gateway command>', this will cause the
gateway to run on CPU0.  The test command can then be prefixed with
'taskset 0xF7 <test command>', so that webthing instances are running
on CPUs 1-7.  This ensures that the gateway doesn't have to fight with
the test harness for CPU time.