Fix test_5_property_change_at_webthing
Test test_5_property_change_at_webthing relied on direct access to the
testthing object to initiate the organic change at the webthing end.

When switching the test harness from using threaded instances of webthings
to use process instances of webtings, this direct access was of course
broken (although we were still accessing the object in it's
pre-forked state, so everything kind of looked OK, but just got broken
in a subtle way).

This change modifies the testthing instance so that it has an additional
'command' message queue which can be used to trigger property changes
directly, and modifies test_5_property_change_at_webthing to use that
queue.